Lobbyをコンポーネント分割
